More finished files
Now that the lc generator script is working properly, it can be added here as well together with the bash scripts to run the entire pipeline in either mode.
readme.txt is also added but not 100% complete yet, as it does not yet have the notes on how to make run_daily_check.sh run every day at a given time. This will be added later when I figure out how to do that.